backend/requirements.txt
# ============================================================
# Mukthi Guru — Backend Dependencies
# Compatible with Python 3.12. Colab-compatible.
# ============================================================

# --- Core Framework ---
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0,<1.0
pydantic>=2.10.0,<3.0
pydantic-settings>=2.7.0,<3.0
python-dotenv>=1.0.0,<2.0

# --- LLM Orchestration ---
langchain>=0.3.15,<1.0
langchain-core>=0.3.31,<1.0
langchain-community>=0.3.15,<1.0
langchain-ollama>=0.3.0,<1.0
langgraph>=0.2.60,<1.0

# --- Vector Database ---
qdrant-client>=1.13.0,<2.0

# --- Embeddings & Reranking ---
sentence-transformers>=3.4.0,<4.0

# --- YouTube Ingestion ---
youtube-transcript-api>=0.6.3,<1.0
yt-dlp>=2025.1.0

# --- Image Handling ---
requests>=2.32.0,<3.0
Pillow>=11.0.0,<12.0

# --- RAPTOR Clustering ---
umap-learn>=0.5.7
scikit-learn>=1.6.0,<2.0

# --- Testing ---
pytest>=8.3.0,<9.0
pytest-asyncio>=0.25.0,<1.0
httpx>=0.28.0,<1.0

# --- Heavy / Optional (install separately if needed) ---
# openai-whisper>=20240930      # Whisper transcription (needs ffmpeg)
# easyocr>=1.7.2                # OCR (large torch dependency)
# nemoguardrails>=0.11.0        # NeMo guardrails (optional safety layer)

# --- Colab Extras (install separately on Colab) ---
# pyngrok==7.2.2
